Move processor generic code to the main Aztec project.
